requirements.txt
asgiref==3.3.1
Django==3.1.7
docx2pdf==0.1.7
fpdf==1.7.2
importlib-metadata==1.7.0
lxml==4.6.3
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.3
protobuf==3.15.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
pywin32==227
six==1.15.0
sqlparse==0.4.1
tqdm==4.59.0
zipp==3.4.1
